+++ libs/ui/vegetable-form/README.md
@@ -1,0 +1,7 @@
+# vegetable-form
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test vegetable-form` to execute the unit tests.
