+++ libs/ui/vegetable-form/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/vegetable-form/vegetable-form.component';
