--- src/app/presentation/modules/vegetable-form/index.ts
@@ -1,2 +1,0 @@
-
-export * from './vegetable-form.component';
