+++ tsconfig.base.json
@@ -12,15 +12,13 @@
     "skipLibCheck": true,
     "baseUrl": "./",
     "paths": {
-      "@core/*": ["src/app/core/*"],
       "@data-layer": ["libs/data-layer/src/index.ts"],
-      "@data/*": ["src/app/data/*"],
-      "@shared/models": ["libs/shared/models/src/index.ts"],
       "@state": ["libs/state/src/index.ts"],
+      "@shared/models": ["libs/shared/models/src/index.ts"],
+      "@ui/card": ["libs/ui/card/src/index.ts"],
       "@ui/listbox": ["libs/ui/listbox/src/index.ts"],
       "@ui/toaster": ["libs/ui/toaster/src/index.ts"],
-      "@ui/card": ["libs/ui/card/src/index.ts"],
-      "@ui/*": ["src/app/presentation/*"]
+      "@ui/vegetable-form": ["libs/ui/vegetable-form/src/index.ts"]
     }
   }
 }
